requirements-advertiser-portal.txt
# Advertiser Portal Requirements
# Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.2
django-cors-headers>=4.1.0
django-extensions>=3.2.0

# Database
psycopg2-binary>=2.9.0
redis>=4.5.0
django-redis>=5.2.0

# Authentication and Security
django-oauth-toolkit>=1.6.0
djangorestframework-simplejwt>=5.2.0
cryptography>=41.0.0
bcrypt>=4.0.0

# Background Tasks
celery>=5.3.0
celery-beat>=2.4.0
flower>=2.0.0

# File Storage and Processing
Pillow>=10.0.0
django-storages>=1.13.0
boto3>=1.26.0
python-magic>=0.4.27

# Email Services
django-anymail>=0.22.0
sendgrid>=6.10.0

# Payment Processing
stripe>=5.4.0
paypal-checkout-sdk>=0.0.3

# Analytics and Reporting
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# API and Integration
requests>=2.31.0
httpx>=0.24.0
urllib3>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Validation and Serialization
marshmallow>=3.19.0
pydantic>=2.1.0
jsonschema>=4.17.0

# Caching
django-cacheops>=7.0.0
django-redis-cache>=3.0.0

# Monitoring and Logging
sentry-sdk>=1.28.0
structlog>=23.1.0
django-structlog>=6.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.2.0
faker>=18.9.0

# Development Tools
django-debug-toolbar>=4.1.0
django-silk>=5.0.0
django-extensions>=3.2.0

# Documentation
drf-yasg>=1.21.0
drf-spectacular>=0.26.0
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.4.0
tqdm>=4.65.0
jinja2>=3.1.0

# Performance
gunicorn>=21.2.0
uvicorn>=0.22.0
whitenoise>=6.5.0

# Data Processing
openpyxl>=3.1.0
xlrd>=2.0.0
python-docx>=0.8.11

# Time and Date
arrow>=1.2.3
pendulum>=2.1.2

# Configuration
python-decouple>=3.8
django-environ>=0.10.0

# Webhooks
django-webhooks>=1.2.0

# Rate Limiting
django-ratelimit>=4.0.0

# File Validation
python-magic>=0.4.27
python-magic-bin>=0.4.14

# Geographic Data
geopy>=2.3.0
django-countries>=7.5.1

# Currency and Finance
forex-python>=1.8.0
moneyed>=2.0.0
django-money>=3.4.0

# Image Processing
opencv-python>=4.8.0
scikit-image>=0.21.0

# Machine Learning (for fraud detection and optimization)
scikit-learn>=1.3.0
joblib>=1.3.0

# Data Validation
cerberus>=1.3.4
schema>=0.7.5

# Template Engine
django-template-tags>=1.3.0

# Internationalization
django-modeltranslation>=0.18.0

# API Documentation
drf-spectacular>=0.26.0
drf-spectacular-sidecar>=2023.10.1

# Background Processing
django-background-tasks>=1.2.5

# Health Checks
django-health-check>=3.17.0

# Performance Monitoring
django-silk>=5.0.0
django-debug-toolbar>=4.1.0

# Security
django-axes>=5.5.0
django-ratelimit>=4.0.0
django-secure-admin>=2.1.0

# Database Migrations
django-migration-linter>=4.5.0

# Environment Management
python-decouple>=3.8
